review code.

+++ orion-ops-machine-monitor-agent/src/main/java/com/orion/ops/machine/monitor/entity/vo/NetBandwidthMetricsStatisticVO.java
@@ -26,11 +26,11 @@ public class NetBandwidthMetricsStatisticVO implements BaseMetricsStatisticsEnti
     /**
      * 上行包数
      */
-    private MetricsStatisticsVO<Double> sentPacket;
+    private MetricsStatisticsVO<Long> sentPacket;
 
     /**
      * 下行包数
      */
-    private MetricsStatisticsVO<Double> recvPacket;
+    private MetricsStatisticsVO<Long> recvPacket;
 
 }
